requirements.txt
streamlit==1.28.1
langchain==0.1.0
langchain-community==0.0.8
langgraph==0.0.1
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
openai==1.3.5
